refactor: Move components for different sections of the sheet into separate files
