+++ dev-client/src/screens/LocationScreens/components/SiteDataSection.tsx
@@ -1,0 +1,40 @@
+/*
+ * Copyright © 2024 Technology Matters
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see https://www.gnu.org/licenses/.
+ */
+
+import {useTranslation} from 'react-i18next';
+
+import {
+  Heading,
+  Text,
+} from 'terraso-mobile-client/components/NativeBaseAdapters';
+import {ScreenContentSection} from 'terraso-mobile-client/components/content/ScreenContentSection';
+
+export const SiteDataSection = () => {
+  const {t} = useTranslation();
+
+  return (
+    <ScreenContentSection title={t('site.soil_id.site_data.title')}>
+      <Text variant="body1">{t('site.soil_id.site_data.description')}</Text>
+      <Heading variant="h6" pt="lg">
+        {t('site.soil_id.site_data.slope.title')}
+      </Heading>
+      <Heading variant="h6" pt="lg">
+        {t('site.soil_id.site_data.soil_properties.title')}
+      </Heading>
+    </ScreenContentSection>
+  );
+};
